+++ src/brand/k.ts
@@ -1,0 +1,14 @@
+import kSpriteO from "../../assets/brand/k-o.png";
+import kSprite from "../../assets/brand/k.png";
+import type { SpriteCrewItem } from "../types/crew";
+
+export const kData: SpriteCrewItem = {
+    name: "k",
+    description: "The first letter of Kaboom",
+    author: "lajbel",
+    secret:
+        "Before this design, there was some others, like a balloom or a bomb",
+    category: "brand",
+    sprite: kSprite,
+    outlined: kSpriteO,
+};