+++ src/brand/ka.ts
@@ -1,0 +1,13 @@
+import kaSpriteO from "../../assets/brand/ka-o.png";
+import kaSprite from "../../assets/brand/ka.png";
+import type { BrandCrewItem } from "../types/crew";
+
+export const kaData: BrandCrewItem = {
+    name: "KA",
+    description: "The K of KAPLAY",
+    author: "misanthrope",
+    secret: "KAPLAY it's a play between KA and PLAY",
+    category: "brand",
+    sprite: kaSprite,
+    outlined: kaSpriteO,
+};