+++ src/brand/kaplayjs.ts
@@ -1,0 +1,13 @@
+import kaplayjsSpriteO from "../../assets/brand/kaplayjs-o.png";
+import kaplayjsSprite from "../../assets/brand/kaplayjs.png";
+import type { SpriteCrewItem } from "../types/crew";
+
+export const kaplayjsData: SpriteCrewItem = {
+    name: "kaplayjs",
+    description: "The fun game library for HTML5 games",
+    author: "misanthrope",
+    secret: "KaPlay!",
+    category: "brand",
+    sprite: kaplayjsSprite,
+    outlined: kaplayjsSpriteO,
+};