+++ src/crew/sukomi.ts
@@ -3,10 +3,10 @@ import sukomiSprite from "../../assets/crew/sukomi.png";
 import type { SpriteCrewItem } from "../types/crew";
 
 export const sukomiData: SpriteCrewItem = {
-    name: "sukomi",
+    name: "Sukomi",
     description: "The best friend of Bobo! Addicted to math.",
     author: "lajbel",
-    secret: "MF's fish. RIP Sukomi",
+    secret: "MF's fish",
     category: "crew",
     sprite: sukomiSprite,
     outlined: sukomiSpriteO,