+++ src/plugin.ts
@@ -13,29 +13,35 @@ export type FontCrewAsset = keyof typeof fontAssets;
 export type KAPLAYCrewPlugin = {
     loadCrew(
         crew: CrewAsset | `${CrewAsset}-o`,
+        name?: string,
     ): Asset<SpriteData>;
     loadCrewFont(
         font: FontCrewAsset | `${FontCrewAsset}-o`,
+        name?: string,
     ): Asset<FontData | BitmapFontData>;
 };
 
 export const crew: KAPLAYPlugin<KAPLAYCrewPlugin> = (k) => {
     return {
-        loadCrew(crew) {
+        loadCrew(crew, name) {
+            name = name ?? crew;
+
             if (crew.endsWith("-o")) {
                 let ocrew = crew.slice(0, -2);
-                return k.loadSprite(crew, crewAssets[ocrew].outlined);
+                return k.loadSprite(name, crewAssets[ocrew].outlined);
             }
             else {
-                return k.loadSprite(crew, crewAssets[crew].sprite);
+                return k.loadSprite(name, crewAssets[crew].sprite);
             }
         },
-        loadCrewFont(font) {
+        loadCrewFont(font, name) {
+            name = name ?? font;
+
             if (font.endsWith("-o")) {
                 let ofont = font.slice(0, -2);
                 const fontData = fontAssets[ofont as "happy"];
                 return k.loadBitmapFont(
-                    ofont,
+                    name,
                     fontData.outlined,
                     fontData.width_o,
                     fontData.height_o,
@@ -44,7 +50,7 @@ export const crew: KAPLAYPlugin<KAPLAYCrewPlugin> = (k) => {
             else {
                 const fontData = fontAssets[font as "happy"];
                 return k.loadBitmapFont(
-                    font,
+                    name,
                     fontData.sprite,
                     fontData.width,
                     fontData.height,